requirements.txt
absl-py==2.1.0
anyio==4.4.0
APScheduler==3.6.3
astunparse==1.6.3
blinker==1.8.2
cachetools==4.2.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
nltk==3.8.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
protobuf==4.25.3
Pygments==2.18.0
python-telegram-bot==13.7
pytz==2024.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.0
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
